requirements.txt
numpy==1.26.3
Pillow==10.2.0
PyAutoGUI==0.9.54
pytesseract==0.3.10
sounddevice==0.4.6
soundfile==0.12.1
keyboard==0.13.5
pandas==2.1.4
opencv-python==4.8.1.78
requests==2.28.2
PyYAML==6.0.1
ultralytics==8.1.1
matplotlib==3.8.2
seaborn==0.13.1
scipy==1.11.4
setuptools
pywin32
tqdm==4.66.1
psutil==5.9.7
thop==0.1.1-2209072238
